@jonsneyers 2021-10-05
If you want to visually compare before/after:

https://sneyers.info/comparison/before/index.jxl_d1.html
https://sneyers.info/comparison/before/index.jxl_d2.html
https://sneyers.info/comparison/before/index.jxl_d2.7_epf0.html
https://sneyers.info/comparison/before/index.jxl_d4.html

https://sneyers.info/comparison/after/index.jxl_d1.html
https://sneyers.info/comparison/after/index.jxl_d2.html
https://sneyers.info/comparison/after/index.jxl_d2.7_epf0.html
https://sneyers.info/comparison/after/index.jxl_d4.html