@lvandeve 2021-10-05
I can see clear differences in d4, however it's not really a fair comparison since the bpp is different, perhaps some after reports with a distance chosen to have similar bpp as the before cases could be useful to judge the visual quality