@jonsneyers 2021-10-05
For a comparison at the same (low) bitrate:

https://sneyers.info/comparison/before/index.jxl_d4.html
https://sneyers.info/comparison/after/index.jxl_d3.7_epf3.html