@lvandeve 2021-10-05
Comparing d4 before with d3.7_epf3 after: the bpp's here are very similar, and I see less artifacts (such as colored ringing) after. In a few cases some subtle details are less present afterwards, but the removal of artifacts is better.

Comparing d2 before and after: the d2 after may be a bit more blurry, but at this distance level this doesn't cause a big visible difference and here too colored ringing is removed, so looks nice to me, especially given the lower bpp after