@jonsneyers 2021-10-05
> Could we make the change roughly bpp invariant -- then it would be easier to compare the quality.
> 
> Currently quality is quite a bit worse (smoother), but also the sizes are smaller.

I'm not sure how to do that.

To me, up to d2 the quality is comparable between before and after (there are differences, but they kind of cancel out), and then it starts to get worse after (but quality/bpp improves compared to before).

What would be the easiest/best way to get d>2 to increase bpp so it stays closer to what it is now?