@jonsneyers 2021-10-05
Also, note that the relevant change here is just in the deadzone thresholds in enc_groups.cc, and it's probably a good idea to spend some more time on doing something better there, given how much impact this has on density.