@jonsneyers 2021-10-06
Comparing with (corpus-level) bpp matching before/after:

Before:
```
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
------------------------------------------------------------------------------------------------------------
jxl:d0.5        13270  4149533    2.5015178   2.793  26.588   1.26147425   0.39095048  0.977969580024      0
jxl:d1          13270  2678938    1.6149796   3.007  26.193   1.78285432   0.61857257  0.998982057118      0
jxl:d2          13270  1656707    0.9987346   3.089  25.221   3.14321733   1.00033337  0.999067509763      0
jxl:d3          13270  1245691    0.7509564   3.043  26.529   3.81969762   1.30922044  0.983167407819      0
jxl:d4          13270  1006160    0.6065567   2.920  14.430   5.86349869   1.59430804  0.967038255837      0
jxl:d6          13270   725042    0.4370866   2.917  14.350   7.62232113   2.12312500  0.927989578498      0
jxl:d8          13270   579813    0.3495363   2.934  13.218   9.35824680   2.58072979  0.902058812963      0
Aggregate:      13270  1383537    0.8340561   2.956  19.990   3.79294549   1.15647749  0.964567048624      0
```

After:
```
Encoding         kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
---------------------------------------------------------------------------------------------------------------
jxl:d0.472         13270  4144875    2.4987097   2.874  26.602   1.40367877   0.38858209  0.970953846304      0
jxl:d0.94          13270  2673508    1.6117061   3.111  27.935   1.85886359   0.61419952  0.989909133340      0
jxl:d1.87          13270  1653298    0.9966795   3.053  26.039   2.69780731   0.98890155  0.985617869482      0
jxl:d2.78          13270  1245878    0.7510691   3.347  27.047   3.91507530   1.29760469  0.974590771378      0
jxl:d3.7:epf3      13270  1004842    0.6057622   2.969  15.778   5.63188648   1.59743222  0.967664011915      0
jxl:d5.5           13270   724283    0.4366291   3.006  15.275   6.76689005   2.11827186  0.924899103947      0
jxl:d7.46          13270   577347    0.3480497   2.949  14.811  10.01088524   2.65066652  0.922563728735      0
Aggregate:         13270  1381230    0.8326653   3.041  21.110   3.75439754   1.15529374  0.961972976700      0
```

So at default speed, that's a 0.72% bpp*pnorm improvement at d0.5/d0.472, a 0.91% improvement at d1/d0.94, a 1.35% improvement at d2/d1.87, a 0.81% improvement at d3/d2.78, a 0.06% regression at d4/d3.7, a 0.33% improvement at d6/d5.5, and a 2.3% regression at d8/d7.46.
I don't really consider pnorm very reliable at d>3 though.

At slower speed, it looks like this:

Before:
```
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
------------------------------------------------------------------------------------------------------------
jxl:d0.5:8      13270  4370248    2.6345743   0.425  22.554   0.82337618   0.35225850  0.928051188684      0
jxl:d1:8        13270  2617319    1.5778330   0.439  24.042   1.47186911   0.61985069  0.978020862446      0
jxl:d2:8        13270  1545126    0.9314687   0.447  23.976   2.68204546   1.04558914  0.973933570221      0
jxl:d3:8        13270  1143533    0.6893711   0.449  24.708   3.82011271   1.39642941  0.962658079241      0
jxl:d4:8        13270   910118    0.5486585   0.434  13.989   4.91579628   1.72018522  0.943794160894      0
jxl:d6:8        13270   643386    0.3878609   0.415  13.355   7.04024458   2.30491263  0.893985437072      0
jxl:d8:8        13270   513104    0.3093213   0.409  12.665   9.20119476   2.83191734  0.875972237922      0
Aggregate:      13270  1293919    0.7800304   0.431  18.565   3.26484250   1.19983034  0.935904121505      0
```

After:
```
Encoding           kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
-----------------------------------------------------------------------------------------------------------------
jxl:d0.49:8          13270  4346131    2.6200355   0.468  25.870   0.78668928   0.35028674  0.917763719441      0
jxl:d0.96:8          13270  2611485    1.5743160   0.473  25.139   1.54455209   0.61277995  0.964709274624      0
jxl:d1.9:8           13270  1540298    0.9285582   0.483  25.319   2.54350781   1.03472489  0.960802262611      0
jxl:d2.85:8          13270  1132572    0.6827633   0.503  25.324   4.01355219   1.39427336  0.951958727094      0
jxl:d3.8:epf3:8      13270   902883    0.5442969   0.472  15.945   5.37939930   1.73230829  0.942890001796      0
jxl:d5.6:8           13270   643187    0.3877409   0.440  14.192   7.13568640   2.31264513  0.896707133662      0
jxl:d7.6:8           13270   509472    0.3071317   0.440  14.353   9.62614918   2.90394983  0.891895142622      0
Aggregate:           13270  1287296    0.7760379   0.468  20.162   3.33447242   1.20093020  0.931967363918      0
```

That's a 1.11% improvement at d0.5/d0.49, a 1.36% improvement at d1/d0.96, a 1.35% improvement at d2/d1.9, a 1.12% improvement at d3/d2.85, a 0.10% improvement at d4/d3.8, a 0.30% regression at d6/d5.6, and a 1.81% regression at d8/d7.6.

At faster speed, the situation is as follows:

Before:
```
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
------------------------------------------------------------------------------------------------------------
jxl:d0.5:6      13270  4145352    2.4989973   3.909  25.738   1.40471935   0.39751908  0.993399100677      0
jxl:d1:6        13270  2669631    1.6093689   4.354  28.066   1.78735065   0.62486282  1.005634796426      0
jxl:d2:6        13270  1650662    0.9950904   4.115  24.518   3.08057809   1.00276867  0.997845448152      0
jxl:d3:6        13270  1237692    0.7461342   4.505  25.530   3.84514308   1.31766203  0.983152721781      0
jxl:d4:6        13270   998583    0.6019890   4.212  14.693   5.31905365   1.60527906  0.966360298719      0
jxl:d6:6        13270   717511    0.4325466   4.343  14.465   7.45280504   2.14149141  0.926294892762      0
jxl:d8:6        13270   572284    0.3449975   4.151  13.761   9.50879097   2.61006500  0.900465953733      0
Aggregate:      13270  1374543    0.8286337   4.223  20.093   3.78893924   1.16683482  0.966878705111      0
```

After:
```
Encoding            kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
------------------------------------------------------------------------------------------------------------------
jxl:d0.471:6          13270  4140697    2.4961910   3.762  23.852   1.34665155   0.39422914  0.984071242263      0
jxl:d0.935:6          13270  2666942    1.6077479   4.297  28.016   1.84220552   0.61946132  0.995937609179      0
jxl:d1.86:6           13270  1650039    0.9947148   4.222  25.385   2.90758896   0.99166471  0.986423558948      0
jxl:d2.78:6           13270  1237316    0.7459075   4.176  25.374   3.91918468   1.30714882  0.975012165296      0
jxl:d3.69:epf3:6      13270   997107    0.6010992   4.221  15.657   5.59986734   1.60972632  0.967605166605      0
jxl:d5.47:6           13270   716588    0.4319902   4.196  14.666   7.42241955   2.13959662  0.924284785533      0
jxl:d7.4:6            13270   571743    0.3446714   4.123  13.187   9.88276958   2.66199714  0.917514232302      0
Aggregate:            13270  1373262    0.8278617   4.139  20.052   3.80805685   1.16441109  0.963971366091      0
```

That's a 0.94% improvement at d0.5/d0.471, a 0.97% improvement at d1/d0.935, a 1.14% improvement at d2/d1.86, a 0.83% improvement at d3/d2.78, a 0.13% regression at d4/d3.69, a 0.22% improvement at d6/d5.47, and a 1.89% regression at d8/d7.4.

It looks to me that for d0.5 to d3, this is a ~1% improvement according to pnorm. This is also the range where I trust pnorm the most (d0.5-d2 is kind of the range it was originally designed for).

For higher distances (d4, d6, d8), it's a bit of a regression according to pnorm, but the question is if pnorm is to be trusted there. To me, having less color ringing and fewer visible DCT base patterns is a visual improvement at those distances, not a regression like pnorm is saying.

At d1 and d2, in my opinion, even without adjusting for bpp, the before/after look about the same in fidelity, even though the after pnorm is a bit worse (I didn't look at d0.5 because it's too hard for me to see anything at that fidelity point). Note that at default and kitten speed, d2 after has slightly worse pnorm than d2 before, but the max BA is slightly better too. So in my opinion, at d0.5 to d2, this PR actually does give a ~4-5% improvement.

One more comparison, this time using wombat with epf0 around the fidelity range I consider most relevant for the web, and matching up according to pnorm instead of bpp:

Before:
```
Encoding           kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
-----------------------------------------------------------------------------------------------------------------
jxl:d1.7:6:epf0      13270  1845490    1.1125411   4.258  29.344   2.48481154   0.89959141  1.000832422788      0
jxl:d2.0:6:epf0      13270  1634327    0.9852429   4.374  31.913   3.08055639   1.01053276  0.995620251040      0
jxl:d2.3:6:epf0      13270  1469133    0.8856568   4.303  30.580   3.14892411   1.11501430  0.987520047681      0
jxl:d2.7:6:epf0      13270  1324166    0.7982645   4.075  30.539   3.55734873   1.23730905  0.987699858756      0
Aggregate:           13270  1556373    0.9382486   4.251  30.581   3.04300604   1.05825074  0.992902262101      0
```

After:
```
Encoding            kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
------------------------------------------------------------------------------------------------------------------
jxl:d1.62:epf0        13270  1822862    1.0989000   3.056  32.982   2.49193501   0.89574618  0.984335453287      0
jxl:d1.89:6:epf0      13270  1614394    0.9732265   4.756  33.914   2.92109323   1.00880156  0.981792367734      0
jxl:d2.15:6:epf0      13270  1461330    0.8809529   4.500  32.924   3.32561851   1.10560118  0.973982524435      0
jxl:d2.55:6:epf0      13270  1305749    0.7871619   4.386  33.992   3.63379288   1.23371950  0.971137000475      0
Aggregate:            13270  1539369    0.9279983   4.116  33.449   3.06251823   1.05366222  0.977796799163      0
```

So here, to reach the same pnorm, 1.37 to 1.68% fewer bytes are needed.